Add benchmark that removes all tasks from a Core
